% Generated by roxygen2 (4.0.2): do not edit by hand
\name{conv_len}
\alias{conv_len}
\title{Convert lengths}
\usage{
conv_len(lengths)
}
\arguments{
\item{\strong{lengths}}{character vector of lengths}
}
\description{
Convert lengths from character vector, containing values like "3/4" and "hd"
into a numeric vector.
}
\details{
This function makes some assumptions, so may need tweaking over time.
For each string in a vector it is first split up by the spaces in it, so "1 3/4"
is split into "1" "3/4", if there are no spaces then it assesses the single
number/letters against a list of common length margins.  If the split vector
has length 2 then it assumes the first element is a whole number, and the
second is a fraction (like "3/4"), it calculates the total
}
\examples{
\dontrun{
lengths <- c("0", "nse", "hd", "3/4", "1 1/2")
conv_len(lengths = lengths)
}
}

