% Generated by roxygen2 (4.0.2): do not edit by hand
\name{conv_margins}
\alias{conv_margins}
\title{Convert lengths beaten into seconds}
\usage{
conv_margins(btn_l, cum_l = TRUE, win_time = NULL, conditions = NULL)
}
\arguments{
\item{\strong{btn_l}}{beaten lengths}

\item{\strong{cum_l}}{default \code{TRUE}, indicates whether \strong{btn_l}
is cumulative or not}

\item{\strong{win_time}}{winners time, optional param, entering will mean
the function returns an individual race time for each runner}

\item{\strong{conditions}}{default \code{NULL}, enter race conditions if BHA
lengths per second scale is desired, or enter a scale manually}
}
\value{
\strong{margins}, in seconds, if \strong{win_time} is entered then
individual times per runner are returned
}
\description{
Converts margins, between the winner of a race and the horses behind, from
lengths into seconds.
}
\details{
As default, the lengths per second scale used is the US convention
of 5 lengths per second.  However, if the race was run under the jurisdiction
of the BHA, then entering an abbrevation of the races conditions will apply
the lengths per second scale employed by the BHA.  See also \link{bha_ls} and
\link{bhascale}.
}
\examples{
\dontrun{
btn_l <- c(0, 0.2, 0.5, 1, 1.5)
conv_margins(btn_l = btn_l)
conv_margins(btn_l = btn_l, cum_l = FALSE)
conv_margins(btn_l = btn_l, win_time = 59)
conv_margins(btn_l = btn_l, win_time = 59, conditions = "f-ggf")
}
}

