% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_times.R
\name{conv_times}
\alias{conv_times}
\title{Convert race times}
\usage{
conv_times(times, regex = NULL)
}
\arguments{
\item{\strong{times}}{times to be converted}

\item{\strong{regex}}{specific regular expression to split times by (if
function doesn't cater for the times entered)}
}
\description{
Convert race times into more workable format.  Most Horse Racing times are
run in minutes, so times this function works on will be recorded as
'minutes:seconds:milliseconds' and 'seconds:milliseconds', which will be
converted into seconds:milliseconds.
}
\details{
Takes a character string representing a race time, splits it up
according to a regular expression that looks for punctuation, a space or the
letter 'm' (full regex = "[[:punct:]]\\s?|\\s?[[:alpha:]]\\s?|\\s+"), so times like '1m39.99',
'1 39.99' or '1.39.99' are split into '1', '39' and '99', it then recombines
to return a numeric time in seconds.  If there are any race times that this
doesn't cater for then use the \strong{regex} param to enter a specific
regular expression to split the times by.
}
\examples{
\dontrun{
times <- c("1.39.99", "1m39.99", "1 39.99")
conv_times(times = times)
}
}

