% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_weight.R
\name{conv_wgt}
\alias{conv_wgt}
\title{Convert weights}
\usage{
conv_wgt(wgt, regex = NULL, from = "st", round = 0)
}
\arguments{
\item{\strong{wgt}}{vector of weights to be converted, see examples.}

\item{\strong{regex}}{specific regular expression to split wgt by (if wgt is
a character string representing stones and lbs)}

\item{\strong{from}}{convert from "kg" to lbs, if wgt is numeric}

\item{\strong{round}}{round weight to nearest}
}
\description{
Convert weight from different formats/measurements into lbs
}
\details{
Converts a numeric weight from kilograms ("kg") or stones ("st")
into lbs.  It also converts a character string of stones and lbs, ie. 9-7
representing 9 stone 7 lbs, into lbs using regex.
}
\examples{
\dontrun{
wgts <- c("9-7", "9-0", "11-8")
conv_wgt(wgt = wgts)

wgts <- c(57, 58, 59)
conv_wgt(wgt = wgts, from = "kg")
}
}

