% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_at_wgts.R
\name{diff_at_wgts}
\alias{diff_at_wgts}
\title{Difference at the weights}
\usage{
diff_at_wgts(btn_lbs, wgt_carried)
}
\arguments{
\item{btn_lbs}{beaten lbs between runners}

\item{wgt_carried}{weight carried (in lbs)}
}
\description{
Calculates the difference at the weights, taking into account the number of
lbs behind the winner the horse was, and the weight the horse carried
compared to the winner.
}

