% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{example_race}
\alias{example_race}
\title{Example Dataset}
\description{
Dataset of two fictional races, to highlight use of a few functions
}
\section{Variables}{

\itemize{
     \item \strong{date} race date
     \item \strong{time} race start time
     \item \strong{pos} finishing position
     \item \strong{horse} horse name
     \item \strong{wintime} winners final time
     \item \strong{going} official going
     \item \strong{code} race code, flat or national hunt race
     \item \strong{btn_l} beaten lengths
     \item \strong{dist} race distance
}
}
\keyword{dataset}

