% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/finishing_speeds.R
\name{fin_sect}
\alias{fin_sect}
\title{Finishing Sectional Time}
\usage{
fin_sect(fin_time, dist, fin_spd, sect_dist)
}
\arguments{
\item{\strong{fin_time}}{final time}

\item{\strong{dist}}{distance}

\item{\strong{fin_spd}}{finishin speed percentage}

\item{\strong{sect_dist}}{closing sectional distance}
}
\description{
Calculates the finishing sectional, reverse engineers a Finishing Speed
Percentage (see \link{fin_spd}) into a sectional time
}
\details{
The two distance parameters, \strong{dist} and \strong{sect_dist},
must be in the same units, be that furlongs, miles, metres, feet, etc.  While
the \strong{fin_time} parameter should be in seconds, and the \strong{fin_spd}
should be a number around 100.
}
\examples{
\dontrun{
# 5 furlong race run in 59 seconds
# finishing speed of 102.61 over closing sectional of 2 furlongs
fin_sect(fin_time = 59, dist = 5, fin_spd = 102.61, sect_dist = 2)
## 23
}
}

