% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finishing_speeds.R
\name{fin_spd}
\alias{fin_spd}
\title{Finishing Speed Percentage}
\source{
Two Simon Rowlands articles explaining Finshing
Speeds (and sectionals):
\href{https://www.timeform.com/Racing/Articles/An_introduction_to_Sectional_Timing_Part_1}{Sectional Timing Part 1}
\href{https://www.timeform.com/Racing/Articles/An_introduction_to_Sectional_Timing_Part_2}{Sectional Timing Part 2}
}
\usage{
fin_spd(fin_time, dist, sect_time, sect_dist)
}
\arguments{
\item{fin_time}{final time}

\item{dist}{race distance}

\item{sect_time}{closing sectional time}

\item{sect_dist}{closing sectional distance}
}
\description{
Calculates finishing speed percentages, another innovation by Simon Rowlands
of Timeform
}
\details{
Parameters must be in the same units, so time params \strong{fin_time}
and \strong{sect_time} will be in seconds, while distance params \strong{dist}
and \strong{sect_dist} will both be in furlongs, or both be in feet, or both
be in miles, metres, etc.
}
\examples{
\dontrun{
# 5 furlong race run in 59 seconds
# closing 2 furlongs run in 23 seconds
fin_spd(fin_time = 59, dist = 5, sect_time = 23, sect_dist = 2)
## 102.61
}
}

