% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcappeR.R
\docType{data}
\name{gulfstream}
\alias{gulfstream}
\title{Gulfstream Park Dataset}
\format{2863 rows
17 columns}
\description{
Dataset containing actual races from Gulfstream Park run in 2013 over 6 and
8 furlongs on the dirt.
}
\section{Variables}{

\itemize{
     \item date race date
     \item race race number
     \item race_type type of race, or class of race ("clm",
     "mdn clm")
     \item dist race distance (furlongs)
     \item surf surface (all dirt in this case)
     \item value race value (or purse)
     \item going official going
     \item trainer horse trainer
     \item jockey horse jockey
     \item j_clm jockey claim (any weight, in lbs, taken off horse)
     \item age horse's age
     \item wgt weight carried
     \item gate starting gate
     \item pos finishing position
     \item horse horse
     \item fintime final time
     \item sect_4f time recorded over first 4 furlongs
}
}
\keyword{dataset}

