% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/conv_lengths.R
\name{helper_len}
\alias{helper_len}
\title{Helper function for conv_len}
\usage{
helper_len(length)
}
\arguments{
\item{length}{character string representing a margin in lengths}
}
\description{
Helper function for conv_len
}

