% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{lbs_per_sec}
\alias{lbs_per_sec}
\title{lbs per second scale}
\usage{
lbs_per_sec(dist, surf = "turf")
}
\arguments{
\item{\strong{dist}}{race distance}

\item{\strong{surf}}{race surface, default is "turf", other option "dirt", if
race is on fibresand, use the dirt option, if all-weather use turf}
}
\value{
lbs per second scale
}
\description{
uses a loess model to return a simple lbs per second scale, race
distance the only input.  As such, at the moment, only races between 5
furlongs and 16 furlongs (2 miles) are covered, if a distance outside this
range is entered the value returned will be NA.
}

