% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zipf_hcp.R
\name{plot.rcapper_zipf_hcp}
\alias{plot.rcapper_zipf_hcp}
\title{plot method for detailed return of \link{zipf_hcp}}
\usage{
\method{plot}{rcapper_zipf_hcp}(x, ...)
}
\description{
plot method for detailed return of \link{zipf_hcp}
}

