% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipf_init.R
\name{plot.rcapper_zipf_init}
\alias{plot.rcapper_zipf_init}
\title{plot method for output of \link{zipf_init}}
\usage{
\method{plot}{rcapper_zipf_init}(x, ...)
}
\description{
plot method for output of \link{zipf_init}
}

