% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zipf_hcp.R
\name{print.rcapper_zipf_hcp}
\alias{print.rcapper_zipf_hcp}
\title{print method for detailed return of \link{zipf_hcp}}
\usage{
\method{print}{rcapper_zipf_hcp}(x, ...)
}
\description{
print method for detailed return of \link{zipf_hcp}
}

