% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zipf_init.R
\name{print.rcapper_zipf_init}
\alias{print.rcapper_zipf_init}
\title{print method for output of \link{zipf_init}}
\usage{
\method{print}{rcapper_zipf_init}(x, ...)
}
\description{
print method for output of \link{zipf_init}
}

