% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcappeR.R
\docType{package}
\name{RcappeR}
\alias{RcappeR}
\alias{RcappeR-package}
\title{RcappeR}
\description{
Tools to aid the analysis and handicapping of Thoroughbred Horse Racing
}
\details{
A collection of tools to aid the analysis and handicapping of Thoroughbred
Horse Racing.  The primary tool \link{zipf_init} initialises a handicap
using race standardisation, which establishes the merits of a horses'
performance in a race by comparing it with different, but similar races.
Other tools, \link{zipf_hcp} and \link{zipf_race} help maintain a handicap.
The package also includes various tools for cleaning and preparing data for
analysis, and the use of the primary functions.
}
\author{
Thomas Heslop
}

