% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rnrs_btn.R
\name{rnrs_btn}
\alias{rnrs_btn}
\title{Runners Beaten}
\usage{
rnrs_btn(pos)
}
\arguments{
\item{\strong{pos}}{finishing position}
}
\description{
Calculates the percentage of runners beaten.
}
\details{
It is best used with the power of the \link{dplyr} package, to
calculate the percentage of runners beaten across numerous races
in the same dataframe (see examples)
}
\examples{
\dontrun{
# dataframe consisting of position (pos) variable and starting gate (gate)
# find percentage of runners beaten per starting gate
races \%>\%
     mutate(pct_btn = rnrs_btn(pos)) \%>\%
     group_by(gate) \%>\%
     summarise(pct_btn = mean(pct_btn))
}
}

