% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.rcapper_zipf_hcp}
\alias{summary.rcapper_zipf_hcp}
\title{summary method for detailed return of \link{zipf_hcp}}
\usage{
\method{summary}{rcapper_zipf_hcp}(x, ...)
}
\description{
summary method for detailed return of \link{zipf_hcp}
}

