% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zipf_init.R
\name{summary.rcapper_zipf_init}
\alias{summary.rcapper_zipf_init}
\title{summary method for output of \link{zipf_init}}
\usage{
\method{summary}{rcapper_zipf_init}(x, ...)
}
\description{
summary method for output of \link{zipf_init}
}

