% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/theme_rcapper.R
\name{theme_rcapper}
\alias{theme_rcapper}
\title{Custom RcappeR ggplot2 theme}
\usage{
theme_rcapper(base_size = 10, base_family = "", ...)
}
\description{
A custom theme to help generate smart and identifiable plots
}

