% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zipf_hcp.R
\name{zipf_hcp}
\alias{zipf_hcp}
\title{Handicap a race using many races}
\usage{
zipf_hcp(race, past_races, race_id, btn_var, rating = NULL,
  results = "detail", .progress = "none")
}
\arguments{
\item{race}{dataframe of race to handicap}

\item{past_races}{dataframe of past races used to handicap}

\item{race_id}{name of variable to split \strong{past_races} up by
so each split is one race}

\item{btn_var}{name of variable in \strong{race} with margins
between horses}

\item{rating}{name of variable (if applicable) in
\strong{past_races} that contains the ratings of those runners}

\item{results}{default \strong{detail}, determines the output, other
option is \strong{simple}, which will return the mean rating.}

\item{.progress}{plyr's progress bar (default = "none", options inc.
"text", "time", "tk" or "win")}
}
\value{
Depending on param in \strong{results} it either returns a solitary
rating (the mean of all ratings returned), or a detailed list that can be
analysed further, the list contains
\itemize{
     \item \strong{n_races} number of races used
     \item \strong{mean_rtg} mean rating for the race
     \item \strong{summary} summary statistics of ratings returned
     \item \strong{ratings} ratings dataframe (with vars \emph{race_id} and
     \emph{zipf_rtg})
}
}
\description{
Assess and establish the performance of a winner by using the ratings from
many different, but similar, races, calls \link{zipf_race}.
}
\details{
The \strong{past_races} dataframe is split according to
\strong{race_id}, so each split should be a small dataframe of a single race.
For each of these races they are entered into the \link{zipf_race} function
as \emph{race_2}, while the \strong{race} being handicapped is the \emph{race}
param.  The return is a dataframe with the \strong{race_id} and a rating for
the winner of \strong{race} based on the data from each race in
\strong{past_races}.
}

