% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zipf_race.R
\name{zipf_race}
\alias{zipf_race}
\title{Handicap a race using one race}
\source{
Article by Simon Rowlands explaining use of Zipf's Law:

\url{https://betting.betfair.com/horse-racing/bloggers/simon-rowlands/simon-rowlands-on-handicapping-060710.html}
}
\usage{
zipf_race(race, btn_var, race_2, rating = NULL)
}
\arguments{
\item{race}{dataframe of race to handicap}

\item{btn_var}{name of variable which contains the margins between
the horses}

\item{race_2}{dataframe of a race to be used to handicap
\strong{race}}

\item{rating}{name of ratings variable (if applicable) in
\strong{race_2}}
}
\description{
Assess the performance of a winner (ie handicap) of a race using race
standardisation; which uses the performances of runners in a different, but
similar, race for this assessment.  It is called by \link{zipf_hcp} and
\link{zipf_init}.
}
\details{
The method of race standardisation used was first explained by Simon
Rowlands and uses Zipfs Law.  The difference at the weights, from the race
to be handicapped, is applied to the second race, creating a vector of
possible ratings the victor could have achieved.  This is where Zipfs Law
plays its part, as depending on the finishing position, different weights
are assigned to each of the ratings in the vector, placing more significance
on horses towards the front of the field.
}

