% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcappeR.R
\docType{data}
\name{bhascale}
\alias{bhascale}
\title{BHA lengths per second scale}
\source{
\href{http://www.britishhorseracing.com/wp-content/uploads/2014/04/Lengths-Per-Second-Scale-Tables.pdf}{BHA Scale PDF}
}
\description{
Dataset containing the lengths per second scale employed by the BHA
}
\section{Variables}{

\itemize{
     \item \strong{race_type} type of race, flat (f) or national hunt (nh)
     \item \strong{going} official going
     \item \strong{scale} lengths per second scale
     \item \strong{conditions} abbreviated concatenation of \strong{race_type}
     and \strong{going}
}
}
\keyword{dataset}

