% Generated by roxygen2 (4.0.2): do not edit by hand
\name{btn_sec}
\alias{btn_sec}
\title{Beaten seconds}
\usage{
btn_sec(times)
}
\arguments{
\item{\strong{times}}{time recorded by each runner}
}
\description{
Calculates the margins (in seconds) between the winner of a race and horses
who finished behind
}

