% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcappeR.R
\docType{data}
\name{example_race}
\alias{example_race}
\title{Example Dataset}
\description{
Dataset of two fictional races, to highlight use of a few functions
}
\section{Variables}{

\itemize{
     \item date race date
     \item time race start time
     \item pos finishing position
     \item horse horse name
     \item wintime winners final time
     \item going official going
     \item code race code, flat or national hunt race
     \item btn_l beaten lengths
     \item dist race distance
}
}
\keyword{dataset}

