% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcappeR.R
\docType{data}
\name{gulfstream}
\alias{gulfstream}
\title{Gulfstream Park Dataset}
\format{2863 rows
17 columns}
\description{
Dataset containing actual races from Gulfstream Park run in 2013 over 6 and
8 furlongs on the dirt.
}
\section{Variables}{

\itemize{
     \item \strong{date} race date
     \item \strong{race} race number
     \item \strong{race_type} type of race, or class of race ("clm",
     "mdn clm")
     \item \strong{dist} race distance (furlongs)
     \item \strong{surf} surface (all dirt in this case)
     \item \strong{value} race value (or purse)
     \item \strong{going} official going
     \item \strong{trainer} horse trainer
     \item \strong{jockey} horse jockey
     \item \strong{j_clm} jockey claim (any weight, in lbs, taken off horse)
     \item \strong{age} horse's age
     \item \strong{wgt} weight carried
     \item \strong{gate} starting gate
     \item \strong{pos} finishing position
     \item \strong{horse} horse
     \item \strong{fintime} final time
     \item \strong{sect_4f} time recorded over first 4 furlongs
}
}
\keyword{dataset}

