% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zipf_init.R
\name{merge_zipf_init}
\alias{merge_zipf_init}
\title{Merge zipf_init to dataframe}
\usage{
merge_zipf_init(zipf_list, races, btn_var = NULL)
}
\arguments{
\item{\strong{zipf_list}}{list returned by \link{zipf_init}}

\item{\strong{races}}{dataframe of races}

\item{\strong{btn_var}}{name of variable in \strong{races} with margins
between horses, if entered it will calculate ratings for all losers using the
winners rating and subtracting \strong{btn_var}}
}
\description{
Take the list returned by \link{zipf_init} and the dataframe
entered into the \link{zipf_init} function, and merge them together,
returning a dataframe with ratings per runner.
}

