% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_times.R
\name{mmss_ss}
\alias{mmss_ss}
\title{convert time in mm.ss.00 to ss.00}
\usage{
mmss_ss(time)
}
\arguments{
\item{\strong{time}}{list of length 1 with character vector of length 2 or 3,
which represent minutes, seconds, and milliseconds}
}
\description{
convert time in mm.ss.00 to ss.00
}

