% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipf_hcp.R
\name{summary.rcapper_zipf_hcp}
\alias{summary.rcapper_zipf_hcp}
\title{summary method for detailed return of \link{zipf_hcp}}
\usage{
\method{summary}{rcapper_zipf_hcp}(x, ...)
}
\description{
summary method for detailed return of \link{zipf_hcp}
}

